SONARJAVA-6827: Replace abstract class with utility class to avoid coupling
Replace AbstractAnonymousClassToLambdaCheck with AnonymousClassToLambdaUtils
utility class. Both AnonymousClassShouldBeLambdaCheck (S1604) and
AnonymousClassOnFunctionalInterfaceCheck (S9357) are now standalone checks
that delegate shared logic to the utility class.
